+++ mycodo/mycodo_flask/forms/forms_dashboard.py
@@ -82,9 +82,7 @@ class DashboardGraph(FlaskForm):
     enable_export = BooleanField(lazy_gettext('Enable Export'))
     enable_range = BooleanField(lazy_gettext('Enable Range Selector'))
     enable_graph_shift = BooleanField(lazy_gettext('Enable Graph Shift'))
-    enable_manual_y_axis = BooleanField(lazy_gettext('Manual Y-Axis Min/Max'))
-    y_axis_min = DecimalField(lazy_gettext('Y-Axis Minimum'))
-    y_axis_max = DecimalField(lazy_gettext('Y-Axis Maximum'))
+    enable_manual_y_axis = BooleanField(lazy_gettext('Enable Manual Y-Axis Min/Max'))
     use_custom_colors = BooleanField(lazy_gettext('Enable Custom Colors'))