+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 5.5.20 (Unreleased)
+
+### Bugfixes
+
+ - Add ability to set custom Graph colors for Math measurements
+ 
+
 ## 5.5.19 (2018-02-06)
 
 ### Features
@@ -8,7 +15,6 @@
 ### Bugfixes
 
  - Fix saving Math Humidity options ([#400](https://github.com/kizniche/mycodo/issues/400))
- - Add ability to set custom Graph colors for Math measurements
 
 
 ## 5.5.18 (2018-02-04)